refactor: separate supabase server client
